# doc/openapi.yaml
openapi: 3.0.0
info:
  title: Home Library Service
  version: 1.0.0
  description: A REST API service for managing a home library of music tracks, albums, artists, and user favorites.
servers:
  - url: http://localhost:4000
    description: Development server
paths:
...
  /favs/track/{id}:
    post:
      summary: Add track to favorites
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '201':
          description: Track added to favorites
    delete:
      summary: Remove track from favorites
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Track removed from favorites
  /favs/album/{id}:
    post:
      summary: Add album to favorites
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '201':
          description: Album added to favorites
    delete:
      summary: Remove album from favorites
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Album removed from favorites
  /favs/artist/{id}:
    post:
      summary: Add artist to favorites
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '201':
          description: Artist added to favorites
    delete:
      summary: Remove artist from favorites
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Artist removed from favorites
...